refactor: temporarily rename the Timer component for case change in the filename
